Se agregaron todos los repositorios de los datos que utilizaran las UI , algunos metodos se modificaron porque los metodos no se apegaban a nuestros requerimientos, todosas las clses son metodos sencillos, en espera de retroalimentacion para su mejora

+++ app/src/main/java/LosPrimos/Durango/calculadoragastos/data/daos/CategoriaDao.kt
@@ -12,7 +12,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface CategoriaDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertCategoria(categoria: Categoria): Long
+    suspend fun insertCategoria(categoria: Categoria): Int
 
     @Update
     suspend fun updateCategoria(categoria: Categoria)
